Allow inline CSS per CSP
Otherwise we don't see any styling in development mode. I haven't found
a way to apply this dynamically using Vite. We would like to avoid
allowing this in production, as it weakens the XSS attack protection.

This might either be no longer needed should we integrate Tailwind
(given it doesn't rely on style tags), or otherwise if we switch to a
deployment where we have full-control over HTTP headers and can remove
the <meta> tag altogether.