Decouple primary and secondary adapters
The page as primary adapter must not know anything about any other
adapter, instead it must work with the given port (the interface the
domain has provided). This way we can exchange the underlying, driven
adapter that is providing playbooks, without touching the page at all.

The implementation makes the injection explicit, but we could also use
Vue's provide/inject.