Collect content for PurgeCSS automatically
This way we don't need to adapt postcss.config.js manually whenever we
start using a new component from PrimeVue - this can easily be forgotten
and lead to surprises. Also viceversa: when no longer using a component
we cannot forget to remove it in the config and purging stays most
efficient.

Addresses MIT-88