Harden PurgeCSS content lookup
We had been collecting content based on the import statement such as

import Card from "primevue/card"
       ^^^^      ^^^^^^^^^^^^^^^

The name of the import ("Card") would have had to match what the file is
named in the PrimeVue component ("Card.vue"), but can be freely choosen
by a developer. This changes the lookup to rely solely on the path part
("primevue/card"), which is fix and relyable. It allows developers to
choose a named import to their liking/needs.

Addresses MIT-88