Play back a contract name to model
All of this is pretty rough, we'll want a `PlaybookRepository.update()`
method or similar. Also the test is pretty much tied to the
implementation and thus brittle. We might gain a bit from decomposing
`ContractScreen` a little bit - the updated title could be emitted from
an extracted component in order to detangle the components.

Addresses MIT-29