Implement proof of concept for saving contract
The implementation doesn't take cross-browser support into account yet,
thus only works in Chrome/Edge at this point.

Also testing is a little rough as we need to rely on a lot of mocking.

We're setting a user's desktop directory (so called wellknown directory)
as the start directory for saving the file. This can easily be changed
to say documents (the allowed directories are defined via a narrow
TypeScript property). The suggested filename in the save dialog is
derived from the contract's name (only dealing with whitespaces so far).

Addresses MIT-29