Simplify test to avoid testing router
We shouldn't be testing the router, the subject under test is the side
menu. Now providing the minimal setup that's required to make rendering
the sut work (my first attempt to mock the router altogether didn't
succeed). Visiting a certain route shouldn't be needed either for this
component to render; the test may also have become slightly faster.