Bring back conditional step rendering
This implementation is based on an (extensible) factory which will
dynamically provide the respective component for Vue to render depending
on the type of step. Thus for each step answer type there will need to
be a corresponding component; for making the ContractEdit component
aware of a new type we'll need to extend the factory.

Unfortunately, because types are erased at runtime, and on top Vue using
proxies, we'll need to manually maintain a type as part of the step
classes (can't compare constructor name!).

Needed to refine the core model in terms of Step further along the way.