Unify contract store + repository, saving contract
Both concepts duplicated functionality, and store implemented what the
repository was supposed to become.

For saving a file there is now a similar approach as the loaders, a
simple port instead of a repository (the idea of implementing
repositories to abstract file system access was seemingly wrong).

One outcome is that we're achieving better symmetry in terms of
implementations, a needful prerequisite for simplicity.

For populating in-memory repositories there is now a data directory
containing json representations of our domain. Vite allows to easily
import json files. This may also become useful for testing.

In order to recreate serialized entities it's necessary to be able to
pass an optional id to their constructors.