Remove e2e tests that weren't a requirement
As discussed with UX/product. The now removed e2e tests would imply an
implementation where it is not possible to start a new cloud contract,
as we would always expect to enter the last edited one (think:
singleton).

The plan was to implement a warning upon a user leaving the edit screen
and ask them to save their current work to a file.