Complete ports & adapters architecture
So far Vue components as driving adapters were directly relying on an
outgoing port for loading from/saving to file. That means, in a way, we
have been circling around the domain at the core (though at least we had
achieved some dependency inversion already).

This change addresses this shortcoming: to make the ports & adapters
architecture complete a domain service (also sometimes referred to as
use case) was required. Said domain service implements the port the
driving adapter (a Vue component) depends on, and depends itself on an
outgoing port, so that it can fulfill both loading and saving files, but
doesn't have a direct dependency on an adapter (infrastructure).

As soon as we will need to add validation logic, i.e. any kind of
business logic, the domain service is the place to start. Without it
there wasn't a place to put it. The domain persistence service may also
be the place to implement some sort of autosaving.