Implement jumping to step via side nav
Note that the preview didn't even need to utilize the store for being
updated live...

The other direction remains to be implemented, that is, keeping the side
nav in sync with the current step in the edit view.

Main work here was done by @zechmeister and taken from a branch.
Simplified approach a little bit.

Addresses MIT-69