Sync contract nav with currently edited step
When navigation back and forth through steps in the main contract edit
view the respective module the current step belongs to needs to be
highlighted.

Adding a unit test for the ContractEdit component required me to replace
the Contract's private `#findStepIndex` method, which was causing an
error when rendering the component in the test. Couldn't find out
exactly whether the error relates to usage of Proxies, whether it's a
TypeScript bug, or a binding issue.

Addresses MIT-69