Pull special case handling out of contract screen
In case the given contract id prop matched "cloud-contract" we had to
create a new contract on the fly as a special case, whereas in all other
cases the prop is supposed to be an id we can use to look up a contract
in a repository.

With this change we pull this special case handling one level up to the
routing component, where it serves as kind of a redirect mechanism. It's
still ugly and should not maybe not exist in the first place. But now we
have achieved a consistent url scheme for when editing any contract, no
matter the entry point (open playbook, open contract, start new). This
should also allow to more easily implement reloading later (MIT-128), as
well as adding further built-in contracts to choose from (EVB-IT), while
the simplified screen component in question doesn't need to be changed
any longer. Instead the parent component could offer another UI to
select contracts to start.

Together with the recently introduced localStorage based repository this
uncovered a bug in our Step entity, where the `type` getter wasn't
properly serialized to JSON and as a result serialized contracts
couldn't correctly be deserialized.