Implement dynamic step additions for single choice
By introducing a `path` getter for traversable entities we are able to
dynamically adapt the path of steps a user has to take to finish the
contract. A step itself may itself provide additional steps to its path
given a particular choice. This concept is implemented here for single
choice steps, but should work just as well for multiple or basically any
step that contains additional, potential steps depending on an answer.

Addresses MIT-93