Improve drag and drop E2E tests
Test the real drop interaction! For some reason I couldn't make it work
by loading the contents of the respective fixtures into a buffer to
produce a file for the DataTransfer object as shown:

https://github.com/microsoft/playwright/issues/10667#issuecomment-998397241

Thus using a minimal hardcoded json string (we want to test the drop
interaction here, and nothing else, thus it should be fine).