Streamline E2E step editing tests
In order to avoid testing certain things over and over (navigating
through a contract to reach a certain step) we start utilizing the
Playwright capability to set up localStorage with a given state, in this
case contracts with particular steps we'd like to test. This won't scale
when testing new step types while relying on a single contract to test
with. New tests would need more navigation.

Furthermore the adapted E2E tests are now decoupled from the (evolving)
contract in the data directory, thus are less likely to break when data
changes, i.e. will require less maintenance.

This is going to make tests faster and more focussed on the subject
under test.