Fix rich text with line break within template
If an instance of RichTextAnswerStep had a `produce` template other than
the default, such as "Prefix... ${answer}", an undesired linebreak
occured after the additional text coming from the template followed by
content entered into the rich text editor (== ${answer}).

To fix this it's now assumed that in conjunction with a customized
template the editor content is supposed to become a single sentence
along without a linebreak, with enclosing `<p>` tags being stripped from
the editor content.

We can't just put the prefix into the editor, as this will produce its
own paragraph and thus a line break.

Addresses MIT-171