Fix navigating back to start page
When creating a new contract/playbook, attempting to navigate back using
the back button wouldn't lead to the start page as expected, but would
create yet another contract/playbook. This was due to the fact, we've
implemented a redirect sort-of, but weren't using `replace()` in the vue
router.

Addresses MIT-177