Avoid regression tests in E2E test suite
This should be last resort, unit testing provides much faster feedback.
Also we usually don't need to test for regression in three browser
engines, and had in fact skipped some, but this produced noise in the
E2E log output.

The tradeoff is that we need to spy on a router function (implementation
detail).

Addresses MIT-177